pyproject.toml
[project]
name = "still-on-task-llm"
version = "0.1.0"
description = "Comparing knowledge update methods on task-tuned LLMs"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "peft>=0.10",
    "trl>=0.8",
    "datasets>=2.18",
    "sentence-transformers>=2.6",
    "faiss-cpu>=1.7.4",
    "wandb>=0.16",
    "omegaconf>=2.3",
    "einops>=0.7",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "huggingface-hub>=0.22",
    "tqdm",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "jupyter", "matplotlib", "seaborn"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sot"]

[tool.ruff]
line-length = 100
target-version = "py311"
